requirements.txt
numpy==1.24.3
pandas==2.0.3
matplotlib==3.7.2
scikit-image==0.19.3
scikit-fmm==2022.3.26
scipy==1.10.1
pytorch==1.7.0
torchvision==0.8.1
opencv-python==4.10.0.84
pyyaml==6.0.2
requests==2.32.3
pyqt==5.15.10
rospkg==1.5.1
catkin-pkg==1.0.0
distro==1.9.0
docutils==0.20.1
paretoset==1.2.3
mlflow==2.8.1

# VLA Integration Dependencies
transformers==4.35.2
accelerate==0.21.0
peft==0.5.0
bitsandbytes==0.41.1
pillow==10.0.1
tokenizers==0.14.1
sentencepiece==0.1.99
protobuf==3.20.3
torch-audio==0.13.1
datasets==2.14.6
gradio==3.50.2
deepspeed==0.12.2

# Enhanced ML Dependencies
tensorboard==2.14.1
wandb==0.16.0
einops==0.7.0
timm==0.9.12
torchmetrics==1.2.0

# Production Dependencies
docker==6.1.3
nvidia-ml-py3==7.352.0
psutil==5.9.6
tqdm==4.66.1

# AWS Training Dependencies
boto3==1.35.0
scipy==1.10.1
scikit-learn==1.3.2